// src/models/grocer.model.ts
    export default interface Grocer {
        email_grocer: string;
        name_grocer: string;
        last_name_grocer: string;
        name_store: string;
        city_grocer: string;
        password_grocer: string;
        neighborhood:string,
        street:string;
        number_street:string;
        number_grocer:string;
      }